+++ src/test/java/org/apache/skywalking/banyandb/v1/client/metadata/MeasureMetadataRegistryTest.java
@@ -44,7 +44,7 @@ public class MeasureMetadataRegistryTest extends AbstractBanyanDBClientTest {
                 .addField(Measure.FieldSpec.newIntField("total").compressWithZSTD().encodeWithGorilla().build())
                 .addField(Measure.FieldSpec.newIntField("value").compressWithZSTD().encodeWithGorilla().build())
                 .addField(Measure.FieldSpec.newFloatField("sum").compressWithZSTD().encodeWithGorilla().build())
-                .addIndex(IndexRule.create("scope", IndexRule.IndexType.INVERTED, IndexRule.IndexLocation.SERIES))
+                .addIndex(IndexRule.create("scope", IndexRule.IndexType.INVERTED))
                 .build();
         this.client.define(expectedMeasure);
         Assert.assertTrue(measureRegistry.containsKey("service_cpm_minute"));
@@ -66,7 +66,7 @@ public class MeasureMetadataRegistryTest extends AbstractBanyanDBClientTest {
                 .addField(Measure.FieldSpec.newIntField("total").compressWithZSTD().encodeWithGorilla().build())
                 .addField(Measure.FieldSpec.newIntField("value").compressWithZSTD().encodeWithGorilla().build())
                 .addField(Measure.FieldSpec.newFloatField("sum").compressWithZSTD().encodeWithGorilla().build())
-                .addIndex(IndexRule.create("scope", IndexRule.IndexType.INVERTED, IndexRule.IndexLocation.SERIES))
+                .addIndex(IndexRule.create("scope", IndexRule.IndexType.INVERTED))
                 .build();
         this.client.define(expectedMeasure);
         List<Measure> actualMeasures = new MeasureMetadataRegistry(this.channel).list("sw_metric");
@@ -86,7 +86,7 @@ public class MeasureMetadataRegistryTest extends AbstractBanyanDBClientTest {
                 .addField(Measure.FieldSpec.newIntField("total").compressWithZSTD().encodeWithGorilla().build())
                 .addField(Measure.FieldSpec.newIntField("value").compressWithZSTD().encodeWithGorilla().build())
                 .addField(Measure.FieldSpec.newFloatField("sum").compressWithZSTD().encodeWithGorilla().build())
-                .addIndex(IndexRule.create("scope", IndexRule.IndexType.INVERTED, IndexRule.IndexLocation.SERIES))
+                .addIndex(IndexRule.create("scope", IndexRule.IndexType.INVERTED))
                 .build();
         this.client.define(expectedMeasure);
         boolean deleted = this.client.delete(expectedMeasure);